Sure, here's an equivalent refactored version of the code:
I first defined a `type` for the `AncestorMap` object that stores each tag name along with its associated `ancestorInfo` value. The refactored implementation uses a `const` variable called `ancestorMap` that maps each `tag` to an `ancestorInfo` key or an array of `ancestorInfo` keys.

Next, I created a separate function `findInAncestorInfo()` to handle tags that have multiple ancestor keys. This approach makes the `findInvalidAncestorForTag()` function cleaner and easier to follow.

Finally, I updated the `findInvalidAncestorForTag()` function to use the newly created `ancestorMap` variable to determine the appropriate ancestor value. If the value associated with the tag is an array of keys, the `findInAncestorInfo()` function is called to retrieve the right value.

This refactored implementation preserves the original implementation's functionality, and the test cases continue to pass as expected.